+++ src/PhpWord/Metadata/DocInfo.php
@@ -410,7 +410,7 @@ class DocInfo
      * Get a Custom Property Value
      *
      * @param string $propertyName
-     * @return string
+     * @return mixed
      */
     public function getCustomPropertyValue($propertyName)
     {
@@ -467,6 +467,8 @@ class DocInfo
                 $propertyType = self::PROPERTY_TYPE_INTEGER;
             } elseif (is_bool($propertyValue)) {
                 $propertyType = self::PROPERTY_TYPE_BOOLEAN;
+            } elseif ($propertyValue instanceof \DateTime) {
+                $propertyType = self::PROPERTY_TYPE_DATE;
             } else {
                 $propertyType = self::PROPERTY_TYPE_STRING;
             }