+++ src/PhpWord/Writer/Word2007/Part/Settings.php
@@ -76,7 +76,7 @@ class Settings extends AbstractPart
     {
         if ($settingValue == '') {
             $xmlWriter->writeElement($settingKey);
-        } else {
+        } elseif (is_array($settingValue) && !empty($settingValue)) {
             $xmlWriter->startElement($settingKey);
 
             /** @var array $settingValue Type hint */
@@ -147,13 +147,14 @@ class Settings extends AbstractPart
         $this->setOnOffValue('w:doNotTrackMoves', $documentSettings->hasDoNotTrackMoves());
         $this->setOnOffValue('w:doNotTrackFormatting', $documentSettings->hasDoNotTrackFormatting());
         $this->setOnOffValue('w:evenAndOddHeaders', $documentSettings->hasEvenAndOddHeaders());
+        $this->setOnOffValue('w:updateFields', $documentSettings->hasUpdateFields());
 
         $this->setThemeFontLang($documentSettings->getThemeFontLang());
         $this->setRevisionView($documentSettings->getRevisionView());
         $this->setDocumentProtection($documentSettings->getDocumentProtection());
         $this->setProofState($documentSettings->getProofState());
         $this->setZoom($documentSettings->getZoom());
-        $this->getCompatibility();
+        $this->setCompatibility();
     }
 
     /**
@@ -215,6 +216,7 @@ class Settings extends AbstractPart
     private function setRevisionView(TrackChangesView $trackChangesView = null)
     {
         if ($trackChangesView != null) {
+            $revisionView = array();
             $revisionView['w:markup'] = $trackChangesView->hasMarkup() ? 'true' : 'false';
             $revisionView['w:comments'] = $trackChangesView->hasComments() ? 'true' : 'false';
             $revisionView['w:insDel'] = $trackChangesView->hasInsDel() ? 'true' : 'false';
@@ -258,7 +260,7 @@ class Settings extends AbstractPart
     /**
      * Get compatibility setting.
      */
-    private function getCompatibility()
+    private function setCompatibility()
     {
         $compatibility = $this->getParentWriter()->getPhpWord()->getCompatibility();
         if ($compatibility->getOoxmlVersion() !== null) {